requirements.txt
blinker==1.4
click==7.1.2
dnspython==1.16.0
dominate==2.5.2
eventlet==0.28.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-MQTT==1.0.7
Flask-SocketIO==4.3.1
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
paho-mqtt==1.5.1
python-engineio==3.13.2
python-socketio==4.6.0
six==1.15.0
typing==3.7.4.3
visitor==0.1.3
Werkzeug==1.0.1
